+++ src/galileo/handlers/langchain/async_handler.py
@@ -1,11 +1,12 @@
 import json
 import logging
 import time
-from typing import Any, Literal, Optional
+from typing import Any, Optional
 from uuid import UUID
 
 from galileo import galileo_context
 from galileo.logger import GalileoLogger
+from galileo.schema.handlers import LANGCHAIN_NODE_TYPE, Node
 from galileo.utils.serialization import EventSerializer, convert_to_string_dict, serialize_to_str
 
 _logger = logging.getLogger(__name__)
@@ -24,42 +25,6 @@ except ImportError:
     BaseMessage = object
     LLMResult = object
     AgentFinish = object
-
-
-SPAN_TYPE = Literal["llm", "retriever", "tool", "workflow"]
-LANGCHAIN_NODE_TYPE = Literal["agent", "chain", "chat", "llm", "retriever", "tool"]
-
-
-class Node:
-    """
-    A node in the Langchain trace.
-
-    Attributes
-    ----------
-    node_type : LANGCHAIN_NODE_TYPE
-        The type of node.
-    span_params : dict[str, Any]
-        The parameters for the span that will be created.
-    run_id : UUID
-        The run ID of the span.
-    parent_run_id : Optional[UUID]
-        The run ID of the parent span.
-    children : List[str]
-        List of run_ids for child nodes
-    """
-
-    def __init__(
-        self,
-        node_type: LANGCHAIN_NODE_TYPE,
-        span_params: dict[str, Any],
-        run_id: UUID,
-        parent_run_id: Optional[UUID] = None,
-    ) -> None:
-        self.node_type: LANGCHAIN_NODE_TYPE = node_type
-        self.span_params: dict[str, Any] = span_params
-        self.run_id: UUID = run_id
-        self.parent_run_id: Optional[UUID] = parent_run_id
-        self.children: list[str] = []
 
 
 class GalileoAsyncCallback(AsyncCallbackHandler):
